PR suggestions and fixes
