+++ src/App.js
@@ -1,5 +1,4 @@
 import './App.css';
-import { useFeaturedBanners } from './utils/hooks/useFeaturedBanners';
 import Header from './Components/Header/Header';
 import {
   BrowserRouter as Router,
@@ -11,8 +10,6 @@ import Cart from './Pages/Cart';
 import Footer from './Components/Footer';
 
 function App() {
-  // const { data, isLoading } = useFeaturedBanners();
-  // console.log(data, isLoading);
   
   const getSearchResult = (newSearch) => {
     console.log(newSearch);
