--- src/App.test.js
@@ -1,6 +1,0 @@
-import { render } from '@testing-library/react';
-import App from './App';
-
-test('renders learn react link', () => {
-  render(<App />);
-});
